라이프 사이클 완성 디테일 추가 및 도트 시스템 모듈화ㅓ ㄱㄱ

+++ Assets/Scripts/MessageArchive.cs
@@ -3,8 +3,8 @@ using UnityEngine;
 //화면을 클릭해서 셀을 배치함 배치한 오브젝트와 위치를 전달
 public class Msg_GroundBatchCell : Message
 {
-    public int cellX;
-    public int cellY;
+    public float cellX;
+    public float cellY;
     public GameObject cellObject;
     public eGroundTouchState touchState;
     
@@ -13,7 +13,7 @@ public class Msg_GroundBatchCell : Message
         Add,Delete
     } 
     
-    public Msg_GroundBatchCell(int cellX,int cellY,GameObject cellObject,eGroundTouchState touchState)
+    public Msg_GroundBatchCell(float cellX,float cellY,GameObject cellObject,eGroundTouchState touchState)
     {
         this.cellX = cellX;
         this.cellY = cellY;
